+++ src/test/java/chess/ChessBoardTest.java
@@ -1,6 +1,7 @@
 package chess;
 
 import chess.piece.*;
+import chess.position.Position;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
