+++ src/test/java/chess/PositionTest.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.position.Position;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
